Fixed send function in promio.c
Well kinda... reading response from the burner is now in a loop in case not all the data is
available in one read call. There is still an issue where reading large amounts of data from
the chip will hang at random points as well as some bytes of data being corrupt.